requirements.txt
wheel==0.37.1
PyYAML==6.0.1
click==8.1.7
setuptools==58.0.4
rich==13.7.0
twine